+++ lib/src/search_app_bar_page/search_app_bar_page_stream.dart
@@ -18,14 +18,31 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
   final Color searchAppBarbackgroundColor;
   final Color searchAppBarModeSearchBackgroundColor;
   final Color searchAppBarElementsColor;
+
+  /// [searchAppBarIconConnectyOffAppBarColor] You can change the color of
+  /// [iconConnectyOffAppBar]. By default = Colors.redAccent.
+  final Color searchAppBarIconConnectyOffAppBarColor;
   final String searchAppBarhintText;
   final bool searchAppBarflattenOnSearch;
   final TextCapitalization searchAppBarcapitalization;
   final List<Widget> searchAppBaractions;
   final double searchAppBarelevation;
   final TextInputType searchAppBarKeyboardType;
+
+  /// [magnifyinGlassColor] Changes the color of the magnifying glass.
+  /// Keeps IconTheme color by default.
   final Color magnifyinGlassColor;
+
+  ///[iconConnectyOffAppBar] Appears when the connection status is off.
+  ///There is already a default icon. If you don't want to present a choice
+  ///[hideDefaultConnectyIconOffAppBar] = true; If you want to have a
+  ///custom icon, do [hideDefaultConnectyIconOffAppBar] = true; and set the
+  ///[iconConnectyOffAppBar]`.
   final bool hideDefaultConnectyIconOffAppBar;
+
+  /// [iconConnectyOffAppBar] Displayed on the AppBar when the internet
+  /// connection is switched off.
+  /// It is always the closest to the center.
   final Widget iconConnectyOffAppBar;
 
   ///  [iconConnectyOffAppBar] Aparece quando o status da conexao é off.
@@ -34,12 +51,21 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
 
   /// Parametros para o Scaffold
 
-  ///  [widgetConnecty] Apenas mostra algo quando esta sem conexao e ainda nao
-  ///  tem o primeiro valor da stream. Se a conexao voltar passa a mostrar
-  /// o [widgetWaiting] até apresentar o primeiro dado
+  ///  [widgetOffConnectyWaiting] Apenas mostra algo quando esta sem conexao
+  ///  e ainda nao tem o primeiro valor da stream. Se a conexao voltar
+  ///  passa a mostrar o [widgetWaiting] até apresentar o primeiro dado
   final Widget widgetWaiting;
-  final Widget widgetConnecty;
+  final Widget widgetOffConnectyWaiting;
+
+  /// [widgetError] You can do what you have when there is an error
+  /// in the stream.
   final Widget widgetError;
+
+  /// [searchePageFloaActionButton] , [searchePageFloaActionButton] ,
+  /// [searchePageFloatingActionButtonLocation] ,
+  /// [searchePageFloatingActionButtonAnimator]  ...
+  /// ...
+  /// are passed on to the Scaffold.
   final Widget searchePageFloaActionButton;
   final FloatingActionButtonLocation searchePageFloatingActionButtonLocation;
   final FloatingActionButtonAnimator searchePageFloatingActionButtonAnimator;
@@ -61,6 +87,10 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
   final bool endDrawerEnableOpenDragGesture;
 
   /// Parametros para o SearcherGetController
+
+  /// [initialData] List to be filtered by Search.
+  /// These widgets will not be displayed. [widgetOffConnectyWaiting] and
+  /// [widgetWaiting]
   final List<T> initialData;
 
   /// [listStream] Just pass the Stream and we are already in charge
@@ -82,8 +112,9 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
   /// The list will be filtered by the person.name contains (default) a query.
   final StringFilter<T> stringFilter;
 
-  /// [compareSort] If you want to pass a function to filter the order
-  /// in which your list appears.
+  ///[compareSort] If you want your list to be sorted, pass the function on.
+  /// Example: (Person a, Person b) => a.name.compareTo(b.name),
+  /// This list will be ordered by the object name parameter.
   final Compare<T> compareSort;
 
   const SearchAppBarPageStream({
@@ -95,7 +126,7 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
     this.stringFilter,
     this.compareSort,
     this.filtersType,
-    this.widgetConnecty,
+    this.widgetOffConnectyWaiting,
 
     /// Paramentros do SearchAppBar
     this.searchAppBartitle,
@@ -104,6 +135,7 @@ class SearchAppBarPageStream<T> extends StatefulWidget {
     this.searchAppBarbackgroundColor,
     this.searchAppBarModeSearchBackgroundColor,
     this.searchAppBarElementsColor,
+    this.searchAppBarIconConnectyOffAppBarColor = Colors.redAccent,
     this.searchAppBarhintText,
     this.searchAppBarflattenOnSearch = false,
     this.searchAppBarcapitalization = TextCapitalization.none,
@@ -174,6 +206,8 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
             iconTheme: widget.searchAppBariconTheme,
             backgroundColor: widget.searchAppBarbackgroundColor,
             searchBackgroundColor: widget.searchAppBarModeSearchBackgroundColor,
+            iconConnectyOffAppBarColor:
+                widget.searchAppBarIconConnectyOffAppBarColor,
             searchElementsColor: widget.searchAppBarElementsColor,
             hintText: widget.searchAppBarhintText,
             flattenOnSearch: widget.searchAppBarflattenOnSearch,
@@ -186,7 +220,7 @@ class _SearchAppBarPageStreamState<T> extends State<SearchAppBarPageStream<T>> {
             magnifyinGlassColor: widget.magnifyinGlassColor),
         body: StreamSearchBuilder<T>(
             initialData: widget.initialData,
-            widgetConnecty: widget.widgetConnecty,
+            widgetConnecty: widget.widgetOffConnectyWaiting,
             stream: widget.listStream,
             searcher: _controller,
             listBuilder: widget.listBuilder,