+++ CHANGELOG.md
@@ -1,12 +1,19 @@
+## [0.2.1]
+- Correction in changing [widgetOffConnectyWaiting] and [widgetWaiting]. Correction to intialList 
+within Stream <List> and Internet Connection. Readme and documentation improvements.
+
 ## [0.2.0]
-- Change in the connection verification method. The default language is now English with Portuguese translation.
+- Change in the connection verification method. The default language is now English with Portuguese 
+translation.
 
 ## [0.1.1]
 - Improved performance. Classes have now become constant and the need for streams has decreased.
 
 ## [0.1.0]
 - You can now use your SearchAppBarPageStream on a StatefulWidget and setState.
-- I added a mechanism to verify the connection and it shows a widget [widgetConnecty] in the first data request. Appears when there is no data and the connection is disconnected. By default, there is also a status icon on the application bar. You can change the color, type or not show.
+- I added a mechanism to verify the connection and it shows a widget [widgetConnecty] in the first 
+data request. Appears when there is no data and the connection is disconnected. By default, there is 
+also a status icon on the application bar. You can change the color, type or not show.
 
 ## [0.0.2]
 - Corrections in reading and example.